restructure gis.utils
