+++ hydromt/gis/_gdal_drivers.py
@@ -1,0 +1,39 @@
+GDAL_DRIVER_CODE_MAP = {
+    "asc": "AAIGrid",
+    "blx": "BLX",
+    "bmp": "BMP",
+    "bt": "BT",
+    "dat": "ZMap",
+    "dem": "USGSDEM",
+    "gen": "ADRG",
+    "gif": "GIF",
+    "gpkg": "GPKG",
+    "grd": "NWT_GRD",
+    "gsb": "NTv2",
+    "gtx": "GTX",
+    "hdr": "MFF",
+    "hf2": "HF2",
+    "hgt": "SRTMHGT",
+    "img": "HFA",
+    "jpg": "JPEG",
+    "kro": "KRO",
+    "lcp": "LCP",
+    "mbtiles": "MBTiles",
+    "mpr/mpl": "ILWIS",
+    "ntf": "NITF",
+    "pix": "PCIDSK",
+    "png": "PNG",
+    "pnm": "PNM",
+    "rda": "R",
+    "rgb": "SGI",
+    "rst": "RST",
+    "rsw": "RMF",
+    "sdat": "SAGA",
+    "sqlite": "Rasterlite",
+    "ter": "Terragen",
+    "tif": "GTiff",
+    "vrt": "VRT",
+    "xpm": "XPM",
+    "xyz": "XYZ",
+}
+GDAL_EXT_CODE_MAP = {v: k for k, v in GDAL_DRIVER_CODE_MAP.items()}
