+++ hydromt/gis/flw.py
@@ -9,7 +9,8 @@ import numpy as np
 import pyflwdir
 import xarray as xr
 
-from hydromt.gis import utils
+from hydromt.gis.raster_utils import affine_to_coords
+from hydromt.gis.vector_utils import nearest
 
 logger = logging.getLogger(__name__)
 
@@ -260,7 +261,7 @@ def upscale_flwdir(
     # setup output DataArray
     ftype = flwdir.ftype
     dims = ds.raster.dims
-    coords = utils.affine_to_coords(
+    coords = affine_to_coords(
         flwdir_out.transform,
         flwdir_out.shape,
         x_dim=ds.raster.x_dim,
@@ -411,7 +412,7 @@ def reproject_hydrography_like(
                 crs=crs,
             )
             gdf0["distnc"] = flwdir.distnc.flat[inflow_idxs]
-            gdf0["idx2"], gdf0["dst2"] = utils.nearest(gdf0, gdf_stream)
+            gdf0["idx2"], gdf0["dst2"] = nearest(gdf0, gdf_stream)
             gdf0 = gdf0.sort_values("distnc", ascending=False).drop_duplicates("idx2")
             gdf0["uparea"] = gdf_stream.loc[gdf0["idx2"].values, "uparea"].values
             # set stream uparea to selected inflow cells and calculate total uparea
