+++ hydromt/model/region.py
@@ -15,7 +15,7 @@ from shapely import box
 from hydromt._typing import NoDataStrategy
 from hydromt._typing.type_def import StrPath
 from hydromt.data_catalog import DataCatalog
-from hydromt.gis import utils as gis_utils
+from hydromt.gis import gis_utils
 from hydromt.model.processes.basin_mask import get_basin_geometry
 from hydromt.plugins import PLUGINS
