+++ tests/data_adapter/test_geodataset_adapter.py
@@ -4,7 +4,7 @@ import xarray as xr
 
 from hydromt._typing import SourceMetadata
 from hydromt.data_catalog.adapters.geodataset import GeoDatasetAdapter
-from hydromt.gis.utils import parse_geom_bbox_buffer
+from hydromt.gis.gis_utils import parse_geom_bbox_buffer
 
 
 class TestGeoDatasetAdapter:
